script: Move runtime files into `script/runtime`
It also moves three files into `script/event_loop` since
I initially thought they should be in runtime, but they
are actually tied to the script_thread instead.

Testing: it compiles
