+++ components/script/dom/debugger/debuggerglobalscope.rs
@@ -53,7 +53,7 @@ use crate::dom::types::{
 use crate::dom::webgpu::identityhub::IdentityHub;
 use crate::event_loop::script_thread::with_script_thread;
 use crate::realms::enter_auto_realm;
-use crate::script_runtime::IntroductionType;
+use crate::runtime::script_runtime::IntroductionType;
 
 #[dom_struct]
 /// Global scope for interacting with the devtools Debugger API.
