+++ components/script/dom/html/embedded_content/htmlmediaelement.rs
@@ -108,8 +108,8 @@ use crate::fetch::fetch::{
     FetchCanceller, RequestWithGlobalScope, create_a_potential_cors_request,
 };
 use crate::fetch::network_listener::{self, FetchResponseListener, ResourceTimingListener};
-use crate::microtask::MicrotaskRunnable;
 use crate::realms::enter_auto_realm;
+use crate::runtime::microtask::MicrotaskRunnable;
 use crate::tasks::task_source::SendableTaskSource;
 
 /// A CSS file to style the media controls.
