+++ components/script/dom/testing/testbinding.rs
@@ -58,7 +58,7 @@ use crate::dom::node::Node;
 use crate::dom::promise::Promise;
 use crate::dom::promisenativehandler::{Callback, PromiseNativeHandler};
 use crate::dom::url::URL;
-use crate::timers::OneshotTimerCallback;
+use crate::event_loop::timers::OneshotTimerCallback;
 
 #[dom_struct]
 pub(crate) struct TestBinding {
