+++ components/script/dom/worklet/worklet.rs
@@ -54,10 +54,10 @@ use crate::dom::workletglobalscope::{
     WorkletGlobalScope, WorkletGlobalScopeInit, WorkletGlobalScopeType, WorkletTask,
 };
 use crate::messaging::{CommonScriptMsg, MainThreadScriptMsg, ScriptEventLoopSender};
-use crate::microtask::MicrotaskQueue;
 use crate::modules::script_module::fetch_a_module_script_graph;
 use crate::realms::enter_auto_realm;
-use crate::script_runtime::{IntroductionType, Runtime, ScriptThreadEventCategory};
+use crate::runtime::microtask::MicrotaskQueue;
+use crate::runtime::script_runtime::{IntroductionType, Runtime, ScriptThreadEventCategory};
 use crate::tasks::task_source::TaskSourceName;
 use crate::url::ensure_blob_referenced_by_url_is_kept_alive;
