+++ components/script/event_loop/mod.rs
@@ -2,6 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
+mod devtools;
 pub(crate) mod document_collection;
 pub(crate) mod document_loader;
 mod script_mutation_observers;
@@ -9,3 +10,5 @@ mod script_mutation_observers;
 pub(crate) mod script_thread;
 pub(crate) mod script_window_proxies;
 mod svg_font;
+pub(crate) mod timers;
+pub(crate) mod webdriver_handlers;
