+++ components/script/event_loop/timers.rs
@@ -44,7 +44,7 @@ use crate::dom::types::{Window, WorkerGlobalScope};
 use crate::dom::xmlhttprequest::XHRTimeoutCallback;
 use crate::event_loop::script_thread::ScriptThread;
 use crate::modules::script_module::{ScriptFetchOptions, module_script_from_reference_private};
-use crate::script_runtime::IntroductionType;
+use crate::runtime::script_runtime::IntroductionType;
 use crate::tasks::task_source::SendableTaskSource;
 
 type TimerKey = i32;
